pyproject.toml
[tool.poetry]
name = "evomerge"
version = "0.1.0"
description = "official repository of Evolutionary Optimization of Model Merging Recipes to reproduce the results"
authors = ["Sakana AI"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.38.2"
pillow = "^10.2.0"
datasets = "^2.18.0"
gradio = "^4.21.0"
accelerate = "^0.27.2"
rouge-score = "^0.1.2"
emoji = "^2.10.1"
fugashi = "^1.3.1"
neologdn = "^0.5.2"
unidic-lite = "^1.0.8"
fasttext = "^0.9.2"
einops = "^0.7.0"
sentencepiece = "^0.2.0"
vllm = {git = "https://github.com/SakanaAI/vllm.git", rev = "538aa22a278dc731bc09bb968af2eae909a834c7"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
